Add encapsulation

+++ src/algorithms/min_cut_graph.hpp
@@ -34,9 +34,9 @@ namespace vg {
 
         }; 
 
-        pair<vector<vector<size_t>>, size_t> kargers_min_cut(Graph graph, const int n_iterations, const int seed, int V);
+        pair<vector<vector<size_t>>, size_t> kargers_min_cut(Graph graph, const int n_iterations, const int seed, size_t V);
 
-        pair<vector<vector<size_t>>, size_t> compute_min_cut(Graph graph, const int n_iterations, const int seed, int V);
+        pair<vector<vector<size_t>>, size_t> compute_min_cut(Graph graph, const int n_iterations, const int seed, size_t V);
 
         
     }
